Move Tooltip into lib/components/core
Fix legacy imports
