+++ src/common/dialog/DocumentInformationDialog.svelte
@@ -2,7 +2,7 @@
   import { _ } from "svelte-i18n";
 
   import Button from "@/common/Button.svelte";
-  import Tooltip from "@/common/Tooltip.svelte";
+  import Tooltip from "../lib/components/common/Tooltip.svelte";
   import HtmlEditor from "@/common/HtmlEditor.svelte";
 
   import { metaDialogs } from "./metaDialogs.js";
